* Added -Wextra and -Wshadow compiler options, to enforce extra warnings, and   fixed the code necessary to comply with these flags * Fixed a bug introduced in the 1.0 release, that caused last modified time   not to be reported to callbacks * Changed the semantics of the put data callback, to only make the callback   enough times to get enough data to satisfy the content length * Fixed the error status when curl doesn't get a valid HTTP response from   S3 to be S3StatusConnectionFailed instead of S3StatusHttpErrorUnknown * Fixed a bug where S3_runall_request_context would stop even though a   callback had added another request * libs3.h doesn't need sys/time.h anymore, it just needs sys/select.h * Some minor documentation enhancements

+++ src/bucket.c
@@ -532,7 +532,7 @@ static S3Status listBucketXmlCallback(const char *elementPath,
                          lbData->commonPrefixLens[which] - 1,
                          "%.*s", dataLen, data);
             if (lbData->commonPrefixLens[which] >=
-                sizeof(lbData->commonPrefixes[which])) {
+                (int) sizeof(lbData->commonPrefixes[which])) {
                 return S3StatusXmlParseFailure;
             }
         }
